% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Authentication}
\alias{Authentication}
\title{Authentication Object}
\usage{
Authentication(rules = NULL, providers = NULL)
}
\arguments{
\item{rules}{A list of authentication rules that apply to individual API methods}

\item{providers}{Defines a set of authentication providers that a service supports}
}
\value{
Authentication object
}
\description{
Authentication Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Authentication` defines the authentication configuration for an API.Example for an API targeted for external use:    name: calendar.googleapis.com    authentication:      providers:      - id: google_calendar_auth        jwks_uri: https://www.googleapis.com/oauth2/v1/certs        issuer: https://securetoken.google.com      rules:      - selector: '*'        requirements:          provider_id: google_calendar_auth
}
