% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{AuthenticationRule}
\alias{AuthenticationRule}
\title{AuthenticationRule Object}
\usage{
AuthenticationRule(oauth = NULL, requirements = NULL, selector = NULL,
  allowWithoutCredential = NULL)
}
\arguments{
\item{oauth}{The requirements for OAuth credentials}

\item{requirements}{Requirements for additional authentication providers}

\item{selector}{Selects the methods to which this rule applies}

\item{allowWithoutCredential}{Whether to allow requests without a credential}
}
\value{
AuthenticationRule object
}
\description{
AuthenticationRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Authentication rules for the service.By default, if a method has any authentication requirements, every requestmust include a valid credential matching one of the requirements.It's an error to include more than one kind of credential in a singlerequest.If a method doesn't have any auth requirements, request credentials will beignored.
}
