% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Backend}
\alias{Backend}
\title{Backend Object}
\usage{
Backend(rules = NULL)
}
\arguments{
\item{rules}{A list of API backend rules that apply to individual API methods}
}
\value{
Backend object
}
\description{
Backend Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Backend` defines the backend configuration for a service.
}
