% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{BackendRule}
\alias{BackendRule}
\title{BackendRule Object}
\usage{
BackendRule(address = NULL, selector = NULL, deadline = NULL)
}
\arguments{
\item{address}{The address of the API backend}

\item{selector}{Selects the methods to which this rule applies}

\item{deadline}{The number of seconds to wait for a response from a request}
}
\value{
BackendRule object
}
\description{
BackendRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A backend rule provides configuration for an individual API element.
}
