% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{CustomHttpPattern}
\alias{CustomHttpPattern}
\title{CustomHttpPattern Object}
\usage{
CustomHttpPattern(path = NULL)
}
\arguments{
\item{path}{The path matched by this custom verb}
}
\value{
CustomHttpPattern object
}
\description{
CustomHttpPattern Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A custom pattern is used for defining custom HTTP verb.
}
