% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{DisableServiceRequest}
\alias{DisableServiceRequest}
\title{DisableServiceRequest Object}
\usage{
DisableServiceRequest()
}
\value{
DisableServiceRequest object
}
\description{
DisableServiceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for DisableService method.
}
\seealso{
Other DisableServiceRequest functions: \code{\link{projects.services.disable}}
}
