% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Endpoint}
\alias{Endpoint}
\title{Endpoint Object}
\usage{
Endpoint(name = NULL, apis = NULL, aliases = NULL, features = NULL,
  allowCors = NULL)
}
\arguments{
\item{name}{The canonical name of this endpoint}

\item{apis}{The list of APIs served by this endpoint}

\item{aliases}{DEPRECATED: This field is no longer supported}

\item{features}{The list of features enabled on this endpoint}

\item{allowCors}{Allowing}
}
\value{
Endpoint object
}
\description{
Endpoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Endpoint` describes a network endpoint that serves a set of APIs.A service may expose any number of endpoints, and all endpoints share thesame service configuration, such as quota configuration and monitoringconfiguration.Example service configuration:    name: library-example.googleapis.com    endpoints:      # Below entry makes 'google.example.library.v1.Library'      # API be served from endpoint address library-example.googleapis.com.      # It also allows HTTP OPTIONS calls to be passed to the backend, for      # it to decide whether the subsequent cross-origin request is      # allowed to proceed.    - name: library-example.googleapis.com      allow_cors: true
}
