% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Enum}
\alias{Enum}
\title{Enum Object}
\usage{
Enum(name = NULL, enumvalue = NULL, options = NULL,
  sourceContext = NULL, syntax = NULL)
}
\arguments{
\item{name}{Enum type name}

\item{enumvalue}{Enum value definitions}

\item{options}{Protocol buffer options}

\item{sourceContext}{The source context}

\item{syntax}{The source syntax}
}
\value{
Enum object
}
\description{
Enum Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Enum type definition.
}
