% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Experimental}
\alias{Experimental}
\title{Experimental Object}
\usage{
Experimental(authorization = NULL)
}
\arguments{
\item{authorization}{Authorization configuration}
}
\value{
Experimental object
}
\description{
Experimental Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Experimental service configuration. These configuration options canonly be used by whitelisted users.
}
