% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Http}
\alias{Http}
\title{Http Object}
\usage{
Http(rules = NULL)
}
\arguments{
\item{rules}{A list of HTTP configuration rules that apply to individual API methods}
}
\value{
Http object
}
\description{
Http Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines the HTTP configuration for a service. It contains a list ofHttpRule, each specifying the mapping of an RPC methodto one or more HTTP REST API methods.
}
