% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{LabelDescriptor}
\alias{LabelDescriptor}
\title{LabelDescriptor Object}
\usage{
LabelDescriptor(key = NULL, description = NULL, valueType = NULL)
}
\arguments{
\item{key}{The label key}

\item{description}{A human-readable description for the label}

\item{valueType}{The type of data that can be assigned to the label}
}
\value{
LabelDescriptor object
}
\description{
LabelDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of a label.
}
