% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{ListEnabledServicesResponse}
\alias{ListEnabledServicesResponse}
\title{ListEnabledServicesResponse Object}
\usage{
ListEnabledServicesResponse(services = NULL, nextPageToken = NULL)
}
\arguments{
\item{services}{Services enabled for the specified parent}

\item{nextPageToken}{Token that can be passed to `ListEnabledServices` to resume a paginated}
}
\value{
ListEnabledServicesResponse object
}
\description{
ListEnabledServicesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for `ListEnabledServices` method.
}
