% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{LogDescriptor}
\alias{LogDescriptor}
\title{LogDescriptor Object}
\usage{
LogDescriptor(name = NULL, description = NULL, displayName = NULL,
  labels = NULL)
}
\arguments{
\item{name}{The name of the log}

\item{description}{A human-readable description of this log}

\item{displayName}{The human-readable name for this log}

\item{labels}{The set of labels that are available to describe a specific log entry}
}
\value{
LogDescriptor object
}
\description{
LogDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of a log type. Example in YAML format:    - name: library.googleapis.com/activity_history      description: The history of borrowing and returning library items.      display_name: Activity      labels:      - key: /customer_id        description: Identifier of a library customer
}
