% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Logging}
\alias{Logging}
\title{Logging Object}
\usage{
Logging(consumerDestinations = NULL, producerDestinations = NULL)
}
\arguments{
\item{consumerDestinations}{Logging configurations for sending logs to the consumer project}

\item{producerDestinations}{Logging configurations for sending logs to the producer project}
}
\value{
Logging object
}
\description{
Logging Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Logging configuration of the service.The following example shows how to configure logs to be sent to theproducer and consumer projects. In the example, the `activity_history`log is sent to both the producer and consumer projects, whereas the`purchase_history` log is only sent to the producer project.    monitored_resources:    - type: library.googleapis.com/branch      labels:      - key: /city        description: The city where the library branch is located in.      - key: /name        description: The name of the branch.    logs:    - name: activity_history      labels:      - key: /customer_id    - name: purchase_history    logging:      producer_destinations:      - monitored_resource: library.googleapis.com/branch        logs:        - activity_history        - purchase_history      consumer_destinations:      - monitored_resource: library.googleapis.com/branch        logs:        - activity_history
}
