% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{LoggingDestination}
\alias{LoggingDestination}
\title{LoggingDestination Object}
\usage{
LoggingDestination(logs = NULL, monitoredResource = NULL)
}
\arguments{
\item{logs}{Names of the logs to be sent to this destination}

\item{monitoredResource}{The monitored resource type}
}
\value{
LoggingDestination object
}
\description{
LoggingDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration of a specific logging destination (the producer projector the consumer project).
}
