% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{MediaDownload}
\alias{MediaDownload}
\title{MediaDownload Object}
\usage{
MediaDownload(enabled = NULL, downloadService = NULL)
}
\arguments{
\item{enabled}{Whether download is enabled}

\item{downloadService}{DO NOT USE THIS FIELD UNTIL THIS WARNING IS REMOVED}
}
\value{
MediaDownload object
}
\description{
MediaDownload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Use this only for Scotty Requests. Do not use this for media support usingBytestream, add instead [][google.bytestream.RestByteStream] as an API toyour configuration for Bytestream methods.
}
