% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Method}
\alias{Method}
\title{Method Object}
\usage{
Method(responseTypeUrl = NULL, options = NULL, responseStreaming = NULL,
  name = NULL, requestTypeUrl = NULL, requestStreaming = NULL,
  syntax = NULL)
}
\arguments{
\item{responseTypeUrl}{The URL of the output message type}

\item{options}{Any metadata attached to the method}

\item{responseStreaming}{If true, the response is streamed}

\item{name}{The simple name of this method}

\item{requestTypeUrl}{A URL of the input message type}

\item{requestStreaming}{If true, the request is streamed}

\item{syntax}{The source syntax of this method}
}
\value{
Method object
}
\description{
Method Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Method represents a method of an api.
}
