% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{MetricDescriptor}
\alias{MetricDescriptor}
\title{MetricDescriptor Object}
\usage{
MetricDescriptor(metricKind = NULL, displayName = NULL,
  description = NULL, unit = NULL, labels = NULL, name = NULL,
  type = NULL, valueType = NULL)
}
\arguments{
\item{metricKind}{Whether the metric records instantaneous values, changes to a value, etc}

\item{displayName}{A concise name for the metric, which can be displayed in user interfaces}

\item{description}{A detailed description of the metric, which can be used in documentation}

\item{unit}{The unit in which the metric value is reported}

\item{labels}{The set of labels that can be used to describe a specific}

\item{name}{The resource name of the metric descriptor}

\item{type}{The metric type, including its DNS name prefix}

\item{valueType}{Whether the measurement is an integer, a floating-point number, etc}
}
\value{
MetricDescriptor object
}
\description{
MetricDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a metric type and its schema. Once a metric descriptor is created,deleting or altering it stops data collection and makes the metric type'sexisting data unusable.
}
