% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{OAuthRequirements}
\alias{OAuthRequirements}
\title{OAuthRequirements Object}
\usage{
OAuthRequirements(canonicalScopes = NULL)
}
\arguments{
\item{canonicalScopes}{The list of publicly documented OAuth scopes that are allowed access}
}
\value{
OAuthRequirements object
}
\description{
OAuthRequirements Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
OAuth scopes are a way to define data and permissions on data. For example,there are scopes defined for 'Read-only access to Google Calendar' and'Access to Cloud Platform'. Users can consent to a scope for an application,giving it permission to access that data on their behalf.OAuth scope specifications should be fairly coarse grained; a user will needto see and understand the text description of what your scope means.In most cases: use one or at most two OAuth scopes for an entire family ofproducts. If your product has multiple APIs, you should probably be sharingthe OAuth scope across all of those APIs.When you need finer grained OAuth consent screens: talk with your productmanagement about how developers will use them in practice.Please note that even though each of the canonical scopes is enough for arequest to be accepted and passed to the backend, a request can still faildue to the backend requiring additional scopes or permissions.
}
