% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Option.value}
\alias{Option.value}
\title{Option.value Object}
\usage{
Option.value()
}
\value{
Option.value object
}
\description{
Option.value Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The option's value packed in an Any message. If the value is a primitive,the corresponding wrapper type defined in google/protobuf/wrappers.protoshould be used. If the value is an enum, it should be stored as an int32value using the google.protobuf.Int32Value type.
}
\seealso{
Other Option functions: \code{\link{Option}}
}
