% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Page}
\alias{Page}
\title{Page Object}
\usage{
Page(subpages = NULL, name = NULL, content = NULL)
}
\arguments{
\item{subpages}{Subpages of this page}

\item{name}{The name of the page}

\item{content}{The Markdown content of the page}
}
\value{
Page object
}
\description{
Page Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a documentation page. A page can contain subpages to representnested documentation set structure.
}
