% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_functions.R
\name{projects.services.disable}
\alias{projects.services.disable}
\title{Disable a managed service for a consumer.Operation<response: google.protobuf.Empty>}
\usage{
projects.services.disable(DisableServiceRequest, name)
}
\arguments{
\item{DisableServiceRequest}{The \link{DisableServiceRequest} object to pass to this method}

\item{name}{Name of the consumer and the service to disable for that consumer}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/service.management
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/service.management)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}

Other DisableServiceRequest functions: \code{\link{DisableServiceRequest}}
}
