% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_functions.R
\name{projects.services.enable}
\alias{projects.services.enable}
\title{Enable a managed service for a consumer with the default settings.Operation<response: google.protobuf.Empty>google.rpc.Status errors may contain agoogle.rpc.PreconditionFailure error detail.}
\usage{
projects.services.enable(EnableServiceRequest, name)
}
\arguments{
\item{EnableServiceRequest}{The \link{EnableServiceRequest} object to pass to this method}

\item{name}{Name of the consumer and the service to enable for that consumer}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/service.management
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/service.management)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}

Other EnableServiceRequest functions: \code{\link{EnableServiceRequest}}
}
