% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_functions.R
\name{projects.services.list}
\alias{projects.services.list}
\title{List enabled services for the specified consumer.}
\usage{
projects.services.list(parent, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{List enabled services for the specified parent}

\item{pageToken}{Token identifying which result to start with; returned by a previous list}

\item{pageSize}{Requested size of the next page of data}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}
}
