% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{PublishedService}
\alias{PublishedService}
\title{PublishedService Object}
\usage{
PublishedService(service = NULL, name = NULL)
}
\arguments{
\item{service}{The service's published configuration}

\item{name}{The resource name of the service}
}
\value{
PublishedService object
}
\description{
PublishedService Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The published version of a Service that is managed byGoogle Service Management.
}
