% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{SearchServicesResponse}
\alias{SearchServicesResponse}
\title{SearchServicesResponse Object}
\usage{
SearchServicesResponse(services = NULL, nextPageToken = NULL)
}
\arguments{
\item{services}{Services available publicly or available to the authenticated caller}

\item{nextPageToken}{Token that can be passed to `ListAvailableServices` to resume a paginated}
}
\value{
SearchServicesResponse object
}
\description{
SearchServicesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for SearchServices method.
}
