% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Service}
\alias{Service}
\title{Service Object}
\usage{
Service(apis = NULL, logs = NULL, types = NULL, http = NULL,
  backend = NULL, systemParameters = NULL, documentation = NULL,
  logging = NULL, monitoredResources = NULL, context = NULL,
  enums = NULL, id = NULL, usage = NULL, metrics = NULL,
  authentication = NULL, experimental = NULL, control = NULL,
  configVersion = NULL, monitoring = NULL, producerProjectId = NULL,
  systemTypes = NULL, visibility = NULL, name = NULL,
  customError = NULL, title = NULL, endpoints = NULL)
}
\arguments{
\item{apis}{A list of API interfaces exported by this service}

\item{logs}{Defines the logs used by this service}

\item{types}{A list of all proto message types included in this API service}

\item{http}{HTTP configuration}

\item{backend}{API backend configuration}

\item{systemParameters}{System parameter configuration}

\item{documentation}{Additional API documentation}

\item{logging}{Logging configuration}

\item{monitoredResources}{Defines the monitored resources used by this service}

\item{context}{Context configuration}

\item{enums}{A list of all enum types included in this API service}

\item{id}{A unique ID for a specific instance of this message, typically assigned}

\item{usage}{Configuration controlling usage of this service}

\item{metrics}{Defines the metrics used by this service}

\item{authentication}{Auth configuration}

\item{experimental}{Experimental configuration}

\item{control}{Configuration for the service control plane}

\item{configVersion}{The version of the service configuration}

\item{monitoring}{Monitoring configuration}

\item{producerProjectId}{The id of the Google developer project that owns the service}

\item{systemTypes}{A list of all proto message types included in this API service}

\item{visibility}{API visibility configuration}

\item{name}{The DNS address at which this service is available,}

\item{customError}{Custom error configuration}

\item{title}{The product title associated with this service}

\item{endpoints}{Configuration for network endpoints}
}
\value{
Service object
}
\description{
Service Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Service` is the root object of Google service configuration schema. Itdescribes basic information about a service, such as the name and thetitle, and delegates other aspects to sub-sections. Each sub-section iseither a proto message or a repeated proto message that configures aspecific aspect, such as auth. See each proto message definition for details.Example:    type: google.api.Service    config_version: 3    name: calendar.googleapis.com    title: Google Calendar API    apis:    - name: google.calendar.v3.Calendar    authentication:      providers:      - id: google_calendar_auth        jwks_uri: https://www.googleapis.com/oauth2/v1/certs        issuer: https://securetoken.google.com      rules:      - selector: '*'        requirements:          provider_id: google_calendar_auth
}
