% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_functions.R
\name{services.search}
\alias{services.search}
\title{Search available services.When no filter is specified, returns all accessible services. Forauthenticated users, also returns all services the calling user has'servicemanagement.services.bind' permission for.}
\usage{
services.search(pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{pageToken}{Token identifying which result to start with; returned by a previous list}

\item{pageSize}{Requested size of the next page of data}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}
}
