% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_functions.R
\docType{package}
\name{serviceuser_googleAuthR}
\alias{serviceuser_googleAuthR}
\alias{serviceuser_googleAuthR-package}
\title{Google Service User API
Enables services that service consumers want to use on Google Cloud Platform, lists the available or enabled services, or disables services that service consumers no longer use.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:13:03
filename: /Users/mark/dev/R/autoGoogleAPI/googleserviceuserv1.auto/R/serviceuser_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/service.management
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/cloud-platform
}
}
