% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{SystemParameterRule}
\alias{SystemParameterRule}
\title{SystemParameterRule Object}
\usage{
SystemParameterRule(parameters = NULL, selector = NULL)
}
\arguments{
\item{parameters}{Define parameters}

\item{selector}{Selects the methods to which this rule applies}
}
\value{
SystemParameterRule object
}
\description{
SystemParameterRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Define a system parameter rule mapping system parameter definitions tomethods.
}
