% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{SystemParameters}
\alias{SystemParameters}
\title{SystemParameters Object}
\usage{
SystemParameters(rules = NULL)
}
\arguments{
\item{rules}{Define system parameters}
}
\value{
SystemParameters object
}
\description{
SystemParameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
### System parameter configurationA system parameter is a special kind of parameter defined by the APIsystem, not by an individual API. It is typically mapped to an HTTP headerand/or a URL query parameter. This configuration specifies which methodschange the names of the system parameters.
}
