% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceuser_objects.R
\name{Visibility}
\alias{Visibility}
\title{Visibility Object}
\usage{
Visibility(rules = NULL)
}
\arguments{
\item{rules}{A list of visibility rules that apply to individual API elements}
}
\value{
Visibility object
}
\description{
Visibility Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Visibility` defines restrictions for the visibility of serviceelements.  Restrictions are specified using visibility labels(e.g., TRUSTED_TESTER) that are elsewhere linked to users and projects.Users and projects can have access to more than one visibility label. Theeffective visibility for multiple labels is the union of each label'selements, plus any unrestricted elements.If an element and its parents have no restrictions, visibility isunconditionally granted.Example:    visibility:      rules:      - selector: google.calendar.Calendar.EnhancedSearch        restriction: TRUSTED_TESTER      - selector: google.calendar.Calendar.Delegate        restriction: GOOGLE_INTERNALHere, all methods are publicly visible except for the restricted methodsEnhancedSearch and Delegate.
}
